Split BufferedWriter (again), start with tests for it
