+++ Firestore/core/src/firebase/firestore/remote/buffered_writer.h
@@ -1,0 +1,64 @@
+/*
+ * Copyright 2018 Google
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#ifndef FIRESTORE_CORE_SRC_FIREBASE_FIRESTORE_REMOTE_BUFFERED_WRITER_H
+#define FIRESTORE_CORE_SRC_FIREBASE_FIRESTORE_REMOTE_BUFFERED_WRITER_H
+
+#include <functional>
+#include <vector>
+
+#include <grpcpp/support/byte_buffer.h>
+
+namespace firebase {
+namespace firestore {
+namespace remote {
+
+class BufferedWriter {
+ public:
+  using WriteFunction = std::function<void(grpc::ByteBuffer&&)>;
+
+  explicit BufferedWriter(WriteFunction&& write_func);
+
+  void Start();
+  void Stop();
+  void Clear();
+
+  bool empty() const {
+    return buffer_.empty();
+  }
+
+  bool IsStarted() const {
+    return is_started_;
+  }
+
+  void Enqueue(grpc::ByteBuffer&& bytes);
+  void OnSuccessfulWrite();
+
+ private:
+  void TryWrite();
+
+  WriteFunction write_func_;
+
+  std::vector<grpc::ByteBuffer> buffer_;
+  bool has_pending_write_ = false;
+  bool is_started_ = false;
+};
+
+}  // namespace remote
+}  // namespace firestore
+}  // namespace firebase
+
+#endif  // FIRESTORE_CORE_SRC_FIREBASE_FIRESTORE_REMOTE_BUFFERED_WRITER_H
